requirements.txt
Flask==2.2.5
requests==2.28.2
python-dotenv==1.0.0
APScheduler==3.9.1.post1
